+++ xmake/languages/cuda/xmake.lua
@@ -77,6 +77,10 @@ language("cuda")
         ,   "target.links"
         ,   "option.links"
         ,   "platform.links"
+        ,   "config.syslinks"
+        ,   "target.syslinks"
+        ,   "option.syslinks"
+        ,   "platform.syslinks"
         }
     ,   shared =
         {
@@ -90,6 +94,10 @@ language("cuda")
         ,   "target.links"
         ,   "option.links"
         ,   "platform.links"
+        ,   "config.syslinks"
+        ,   "target.syslinks"
+        ,   "option.syslinks"
+        ,   "platform.syslinks"
         }
     ,   static = 
         {
@@ -112,6 +120,7 @@ language("cuda")
 
                 ,   {category = "Cross Complation Configuration/Builtin Flags Configuration"    }
                 ,   {nil, "links",      "kv", nil,          "The Link Libraries"                }
+                ,   {nil, "syslinks",   "kv", nil,          "The System Link Libraries"         }
                 ,   {nil, "linkdirs",   "kv", nil,          "The Link Search Directories"       }
                 ,   {nil, "includedirs","kv", nil,          "The Include Search Directories"    }
                 }