+++ xmake/languages/golang/api.lua
@@ -30,11 +30,13 @@ function apis()
     {
         -- target.add_xxx
         "target.add_links"
+    ,   "target.add_syslinks"
     ,   "target.add_gcflags"
     ,   "target.add_ldflags"
     ,   "target.add_arflags"
         -- option.add_xxx
     ,   "option.add_links"
+    ,   "option.add_syslinks"
     ,   "option.add_gcflags"
     ,   "option.add_ldflags"
     ,   "option.add_arflags"