+++ xmake/languages/golang/xmake.lua
@@ -78,6 +78,10 @@ language("golang")
         ,   "target.links"
         ,   "option.links"
         ,   "platform.links"
+        ,   "config.syslinks"
+        ,   "target.syslinks"
+        ,   "option.syslinks"
+        ,   "platform.syslinks"
         }
     ,   shared =
         {
@@ -91,6 +95,10 @@ language("golang")
         ,   "target.links"
         ,   "option.links"
         ,   "platform.links"
+        ,   "config.syslinks"
+        ,   "target.syslinks"
+        ,   "option.syslinks"
+        ,   "platform.syslinks"
         }
     }
 
@@ -107,6 +115,7 @@ language("golang")
 
                 ,   {category = "Cross Complation Configuration/Builtin Flags Configuration"   }
                 ,   {nil, "links",      "kv", nil,          "The Link Libraries"               }
+                ,   {nil, "syslinks",   "kv", nil,          "The System Link Libraries"        }
                 ,   {nil, "linkdirs",   "kv", nil,          "The Link Search Directories"      }
                 ,   {nil, "includedirs","kv", nil,          "The Include Search Directories"   }
                 }