+++ xmake/languages/swift/xmake.lua
@@ -100,13 +100,17 @@ language("swift")
         ,   "platform.rpathdirs"
         ,   "platform.frameworkdirs"
         ,   "config.links"
-        ,   "config.frameworks"
         ,   "target.links"
-        ,   "target.frameworks"
         ,   "option.links"
-        ,   "option.frameworks"
         ,   "platform.links"
+        ,   "config.frameworks"
+        ,   "target.frameworks"
+        ,   "option.frameworks"
         ,   "platform.frameworks"
+        ,   "config.syslinks"
+        ,   "target.syslinks"
+        ,   "option.syslinks"
+        ,   "platform.syslinks"
         }
     ,   shared =
         {
@@ -123,13 +127,17 @@ language("swift")
         ,   "platform.linkdirs"
         ,   "platform.frameworkdirs"
         ,   "config.links"
-        ,   "config.frameworks"
         ,   "target.links"
-        ,   "target.frameworks"
         ,   "option.links"
-        ,   "option.frameworks"
         ,   "platform.links"
+        ,   "config.frameworks"
+        ,   "target.frameworks"
+        ,   "option.frameworks"
         ,   "platform.frameworks"
+        ,   "config.syslinks"
+        ,   "target.syslinks"
+        ,   "option.syslinks"
+        ,   "platform.syslinks"
         }
     ,   static = 
         {
@@ -151,6 +159,7 @@ language("swift")
 
                 ,   { category = "Cross Complation Configuration/Builtin Flags Configuration"                        }
                 ,   { nil, "links",         "kv", nil,          "The Link Libraries"                                 }
+                ,   { nil, "syslinks",      "kv", nil,          "The System Link Libraries"                          }
                 ,   { nil, "linkdirs",      "kv", nil,          "The Link Search Directories"                        }
                 ,   { nil, "includedirs",   "kv", nil,          "The Include Search Directories"                     }
                 ,   { nil, "frameworks",    "kv", nil,          "The Frameworks"                                     }