+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#233](https://github.com/tboox/xmake/issues/233): Support windres for mingw platform
 * [#239](https://github.com/tboox/xmake/issues/239): Add cparser compiler support
 * Add plugin manager `xmake plugin --help`
+* Add `add_syslinks` api to add system libraries dependence
 
 ### Changes
 
@@ -507,6 +508,7 @@
 * [#233](https://github.com/tboox/xmake/issues/233): 对mingw平台增加windres的支持
 * [#239](https://github.com/tboox/xmake/issues/239): 添加cparser编译器支持
 * 添加插件管理器，`xmake plugin --help`
+* 添加`add_syslinks`接口去设置系统库依赖，分离与`add_links`添加的库依赖之间的链接顺序
 
 ### 改进